requirements.txt
streamlit
streamlit-extras
pandas
Pillow
